Fix RS Member struct using wrong data types

+++ pkg/opsmanager/schema_automation.go
@@ -84,13 +84,13 @@ type Role struct {
 
 // Member configs
 type Member struct {
-	ID          int    `json:"_id"`
-	ArbiterOnly bool   `json:"arbiterOnly"`
-	Hidden      bool   `json:"hidden"`
-	Host        string `json:"host"`
-	Priority    int    `json:"priority"`
-	SlaveDelay  int    `json:"slaveDelay"`
-	Votes       int    `json:"votes"`
+	ID          int     `json:"_id"`
+	ArbiterOnly bool    `json:"arbiterOnly"`
+	Hidden      bool    `json:"hidden"`
+	Host        string  `json:"host"`
+	Priority    float64 `json:"priority"`
+	SlaveDelay  int     `json:"slaveDelay"`
+	Votes       float64 `json:"votes"`
 }
 
 // ReplicaSet configs
